fix faulty test imports due to package refactoring
