+++ src/test/scala/edu/ie3/powerFactory2psdm/model/PreprocessedPfGridModelSpec.scala
@@ -7,7 +7,7 @@
 package edu.ie3.powerFactory2psdm.model
 
 import edu.ie3.powerFactory2psdm.exception.pf.{ConversionException, MissingParameterException}
-import edu.ie3.powerFactory2psdm.model.powerfactory.RawPfGridModel.ProjectSettings
+import edu.ie3.powerFactory2psdm.model.RawPfGridModel.ProjectSettings
 import edu.ie3.powerFactory2psdm.model.setting.ConversionPrefixes
 import org.scalatest.PrivateMethodTester
 import org.scalatest.matchers.should.Matchers
