+++ src/test/scala/edu/ie3/powerFactory2psdm/model/entity/NodeSpec.scala
@@ -7,6 +7,7 @@
 package edu.ie3.powerFactory2psdm.model.entity
 
 import edu.ie3.powerFactory2psdm.exception.pf.MissingParameterException
+import edu.ie3.powerFactory2psdm.model.RawPfGridModel.{ConElms, Nodes}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
