+++ src/test/scala/edu/ie3/powerFactory2psdm/util/GridPreparatorSpec.scala
@@ -6,7 +6,7 @@
 
 package edu.ie3.powerFactory2psdm.util
 
-import edu.ie3.powerFactory2psdm.model.powerfactory.RawPfGridModel.Switches
+import edu.ie3.powerFactory2psdm.model.RawPfGridModel.Switches
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
